Refactor: DTO 패키지 구조 변경 & DTO명 변경 (#21)
