--- src/main/java/tavebalak/OTTify/genre/dto/SecondGenreUpdateRequestDTO.java
@@ -1,14 +1,0 @@
-package tavebalak.OTTify.genre.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotNull;
-
-@Getter
-@NoArgsConstructor
-public class SecondGenreUpdateRequestDTO {
-
-    @NotNull
-    private Long id;
-}
