+++ src/main/java/tavebalak/OTTify/genre/dto/request/GenreUpdateDTO.java
@@ -1,4 +1,4 @@
-package tavebalak.OTTify.genre.dto;
+package tavebalak.OTTify.genre.dto.request;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,7 +7,7 @@ import javax.validation.constraints.NotNull;
 
 @Getter
 @NoArgsConstructor
-public class FirstGenreUpdateRequestDTO {
+public class GenreUpdateDTO {
 
     @NotNull
     private Long id;
