+++ src/main/java/tavebalak/OTTify/user/service/UserService.java
@@ -1,10 +1,9 @@
 package tavebalak.OTTify.user.service;
 
 import tavebalak.OTTify.genre.dto.request.GenreUpdateDTO;
-import tavebalak.OTTify.genre.dto.request.SecondGenreUpdateRequestDTO;
 
 public interface UserService {
 
     Long update1stGenre(Long userId, GenreUpdateDTO updateRequestDTO);
-    Long update2ndGenre(Long userId, SecondGenreUpdateRequestDTO updateRequestDTO);
+    Long update2ndGenre(Long userId, GenreUpdateDTO updateRequestDTO);
 }
